ati-vga: Fix sign extension in DEFAULT_SC_BOTTOM_RIGHT read
Cast default_sc_bottom to uint32_t before shifting left to prevent
sign extension when assigning to uint64_t.

Fixes: CID 1645615
Message-Id: <20260309212206.2457691-1-chad@jablonski.xyz>

+++ hw/display/ati.c
@@ -513,7 +513,7 @@ static uint64_t ati_mm_read(void *opaque, hwaddr addr, unsigned int size)
         val |= s->regs.default_tile << 16;
         break;
     case DEFAULT_SC_BOTTOM_RIGHT:
-        val = (s->regs.default_sc_bottom << 16) |
+        val = ((uint32_t)s->regs.default_sc_bottom << 16) |
               s->regs.default_sc_right;
         break;
     case SC_TOP:
